refactor(core): split executor.ts into cohesive modules with a run context
executor.ts had grown to ~1900 lines mixing orchestration, output,
locking, wait resolution, lifecycle, and the scheduler — flagged by the
repo's one-domain-per-file guideline. Split behaviour-identically:

- Introduce RunContext, bundling the nine per-run values the run functions
  shared, and pass it as a single argument instead of the 11- and
  14-parameter positional signatures runTarget/runForEachTarget/
  runSequential/runScheduled carried. Each destructures the fields it uses,
  so the bodies are unchanged.
- Extract cohesive modules: run_support.ts (RunEnv + outcome types +
  errorMessage), reporter.ts (the Reporter and its wrappers), lifecycle.ts,
  lock.ts (target cross-run lock), wait_resolution.ts (the .waitsFor gate),
  and scheduler.ts (the mutually-recursive target-running + scheduling
  engine). executor.ts is now the orchestrator: parameters, state, cache,
  cancellation, and execute() itself.

executor.ts drops from 1887 to 808 lines. No public API change — Reporter
is re-exported from executor.ts so existing importers are unaffected. All
tests pass unchanged.
