+++ packages/core/src/executor.ts
@@ -13,11 +13,27 @@
  * their shared prerequisite a single time.
  */
 
-import type { Build, BuildResult, TargetStatus } from "./build.ts";
-import { defaultReadEnv, delay, runWithTimeout } from "./internal.ts";
+import type { Build, BuildResult } from "./build.ts";
+import { defaultReadEnv } from "./internal.ts";
+import {
+  consoleReporter,
+  redactingReporter,
+  type Reporter,
+  safeReporter,
+  silentReporter,
+} from "./reporter.ts";
+export type { Reporter } from "./reporter.ts";
+import { makeLifecycle } from "./lifecycle.ts";
+import type { RunEnv, RunOutcome } from "./run_support.ts";
+import {
+  cpuCount,
+  resolveConcurrency,
+  type RunContext,
+  runScheduled,
+  runSequential,
+} from "./scheduler.ts";
 import { discoverTargets, resolveOrderingEdges } from "./build.ts";
 import { type OrderingEdge, planGraph } from "./graph.ts";
-import { withAmbientEcho } from "./ambient_echo.ts";
 import {
   discoverParameters,
   ParameterError,
@@ -38,97 +54,23 @@ import {
 } from "./affected.ts";
 import { type RemoteCacheStore, resolveRemoteStore } from "./remote_cache.ts";
 import { isCI } from "./host.ts";
-import { ServiceBuilder, ServiceRegistry } from "./service.ts";
+import { ServiceRegistry } from "./service.ts";
 import { Redactor } from "./redact.ts";
 import { absolutePath } from "./path.ts";
-import {
-  type ForEachItem,
-  ForEachSettings,
-  type ForEachSpec,
-  LockSettings,
-  type OnTimeout,
-  type Remediation,
-  type TargetBuilder,
-  type TargetContext,
-  WaitSettings,
-} from "./target.ts";
-import type { Configure } from "./tooling.ts";
-import type { WaitContext } from "./wait.ts";
-import type {
-  RunRecord,
-  SignalRecord,
-  WaitDisposition,
-  WaitState,
-} from "./state/types.ts";
+import type { TargetBuilder } from "./target.ts";
+import type { RunRecord, SignalRecord, WaitState } from "./state/types.ts";
 import { withAmbientSignal } from "./ambient_signal.ts";
 import { defaultStateHost, type StateStore } from "./state/store.ts";
 import { resolveStateStore } from "./state/resolve.ts";
 import { buildRunRecord, ciRunUrl, resolveActor } from "./state/record.ts";
-import { inMemoryStateHandle, RunStateWriter } from "./state/writer.ts";
+import { RunStateWriter } from "./state/writer.ts";
 import { cancelEvent, compensationEvents, runCompensations } from "./cancel.ts";
-import { LockConflictError, type LockHolder } from "./state/lock.ts";
-import { parseDuration } from "./duration.ts";
-import type { Plugin, RunInfo, TargetTiming } from "./plugin.ts";
-import {
-  detectWidth,
-  type Style,
-  type TargetReport,
-  targetWaitFooter,
-} from "./report.ts";
+import type { Plugin, RunInfo } from "./plugin.ts";
+import { detectWidth, type Style, type TargetReport } from "./report.ts";
 import { defaultRenderer, type Renderer } from "./renderer.ts";
 
 /** The artifact directory (under the repo root) for the cache store. */
 const ARTIFACT_DIR = ".zuke";
-
-/** Sink for executor output, defaulting to the console. Overridable in tests. */
-export interface Reporter {
-  /** Write an informational line. */
-  info(line: string): void;
-  /** Write an error line. */
-  error(line: string): void;
-}
-
-const consoleReporter: Reporter = {
-  info: (line) => console.log(line),
-  error: (line) => console.error(line),
-};
-
-const silentReporter: Reporter = { info: () => {}, error: () => {} };
-
-/** Wrap a reporter so every line is passed through the {@link Redactor} first. */
-function redactingReporter(inner: Reporter, redactor: Redactor): Reporter {
-  return {
-    info: (line) => inner.info(redactor.redact(line)),
-    error: (line) => inner.error(redactor.redact(line)),
-  };
-}
-
-/**
- * Wrap a reporter so a failing write can never escape into the run. Output is a
- * best-effort side effect: a sink that throws — a custom reporter with a bug, or
- * the default console raising `BrokenPipe`/EPIPE when stdout is piped to a reader
- * that closed early (`zuke build | head`) — must not turn into a rejection that
- * unwinds the scheduler and strands the durable run record `running`. A thrown
- * write is dropped; every other write still lands.
- */
-function safeReporter(inner: Reporter): Reporter {
-  return {
-    info: (line) => {
-      try {
-        inner.info(line);
-      } catch {
-        // best-effort: a broken output sink must not break the build
-      }
-    },
-    error: (line) => {
-      try {
-        inner.error(line);
-      } catch {
-        // best-effort: a broken output sink must not break the build
-      }
-    },
-  };
-}
 
 /** Options for {@link execute}. */
 export interface ExecuteOptions {
@@ -365,52 +307,6 @@ function resolveStyle(options: ExecuteOptions, github: boolean): Style {
   return { github, color, width: detectWidth() };
 }
 
-/**
- * Open a target's section — a collapsible group under GitHub Actions, or a
- * ruled header in a terminal, separated from the previous block by a blank
- * line.
- */
-function openTarget(
-  r: Reporter,
-  renderer: Renderer,
-  style: Style,
-  name: string,
-  opened: number,
-): void {
-  if (!style.github && opened > 0) r.info("");
-  for (const line of renderer.targetHeader(style, name)) r.info(line);
-}
-
-/** Close a target's section after it succeeded. */
-function passTarget(
-  r: Reporter,
-  renderer: Renderer,
-  style: Style,
-  name: string,
-  ms: number,
-): void {
-  for (const line of renderer.targetPassFooter(style, name, ms)) r.info(line);
-}
-
-/** Close a target's section after it failed and surface the error. */
-function failTarget(
-  r: Reporter,
-  renderer: Renderer,
-  style: Style,
-  name: string,
-  ms: number,
-  error: unknown,
-): void {
-  const { info, error: err } = renderer.targetFailFooter(
-    style,
-    name,
-    ms,
-    error,
-  );
-  for (const line of info) r.info(line);
-  for (const line of err) r.error(line);
-}
-
 /** Append the Markdown job-summary table to `GITHUB_STEP_SUMMARY`, if set. */
 function writeJobSummary(
   renderer: Renderer,
@@ -439,972 +335,6 @@ function writeJobSummary(
     // Best-effort: an unwritable summary file must never fail the build.
   }
 }
-
-/** The result of one target, plus the framework error if it failed. */
-interface TargetOutcome {
-  status: TargetStatus;
-  ms: number;
-  error?: unknown;
-  /**
-   * For a `.forEach(...)` fan-out target, the reports of its materialised
-   * sub-targets, surfaced into the build summary and run record beneath the
-   * parent row. Undefined for an ordinary target.
-   */
-  children?: TargetReport[];
-}
-
-/** What a run (sequential or parallel) produced, fed into the shared summary. */
-interface RunOutcome {
-  reports: TargetReport[];
-  executed: string[];
-  failure: unknown;
-  aborted: boolean;
-  /** True when the run parked at a `.waitsFor(...)` gate rather than finishing. */
-  suspended: boolean;
-}
-
-/**
- * Per-run values threaded to the schedulers and each target: the run id, the
- * cancellation signal handed to every {@link TargetContext}, and the optional
- * durable-state writer that records transitions.
- */
-interface RunEnv {
-  runId: string;
-  signal: AbortSignal;
-  writer?: RunStateWriter;
-  /** The resolved state store, if any — needed to acquire cross-run locks. */
-  store?: StateStore;
-  /** The run's actor, stamped on a lock holder. */
-  actor: string;
-  /** A link to this run (CI job), stamped on a lock holder when known. */
-  runUrl?: string;
-  /** External signals received so far, exposed to bodies via `ctx.signals`. */
-  signals: ReadonlyMap<string, SignalRecord>;
-  /** On a resume, target names already succeeded — seeded done, never re-run. */
-  done?: ReadonlySet<string>;
-  /**
-   * On a resume, each still-waiting target's previously recorded {@link WaitState}
-   * — so a re-suspend preserves the original timeout deadline instead of
-   * recomputing `now + timeout` (which would push it forward on every
-   * `resume --check` and mean the timeout never fires).
-   */
-  priorWaits?: ReadonlyMap<string, WaitState>;
-}
-
-/** A failure's message, or `undefined` when there was none — for the state record. */
-function errorMessage(error: unknown): string | undefined {
-  if (error === undefined) return undefined;
-  return error instanceof Error ? error.message : String(error);
-}
-
-/** A held cross-run lock: `release` clears its heartbeat and frees it. */
-interface HeldLock {
-  release(): Promise<void>;
-}
-
-/** The default conflict guidance when a target declares no `onConflict`. */
-function defaultConflictGuidance(key: string, holder: LockHolder): string {
-  const url = holder.runUrl === undefined ? "" : ` — ${holder.runUrl}`;
-  return `Lock "${key}" is held by ${holder.actor} (run ${holder.runId}) ` +
-    `since ${holder.since}${url}. Wait for that run to finish, or stop it, ` +
-    `then retry.`;
-}
-
-/**
- * Acquire a target's cross-run lock, if it declares one. Returns `null` when it
- * declares no lock, or a {@link HeldLock} once acquired. Throws a
- * {@link LockConflictError} when another run holds it, or a friendly error when
- * a lock is declared but no store is configured.
- */
-async function acquireTargetLock(
-  t: TargetBuilder,
-  env: RunEnv,
-): Promise<HeldLock | null> {
-  const configure = t.lock_;
-  if (configure === undefined) return null;
-  // Run the settings lambda now — after parameters have resolved — so a key
-  // built from `this.<param>.value` sees the final value.
-  const settings = configure(new LockSettings());
-  const name = t.name_ ?? "?";
-  const key = settings.key_;
-  if (key === undefined) {
-    throw new Error(
-      `Target "${name}" .lock(...) set no key — call s.lockKey(...) or s.key(...).`,
-    );
-  }
-  const store = env.store;
-  if (store === undefined) {
-    throw new Error(
-      `Target "${name}" declares .lock("${key}") but no state store is ` +
-        `configured — a lock needs one. Pass --state, set ZUKE_STATE_DIR / ` +
-        `ZUKE_STATE_URL, or override stateStore().`,
-    );
-  }
-  if (settings.ttl_ === undefined) {
-    throw new Error(
-      `Target "${name}" .lock("${key}") set no TTL — call s.withTtl(...).`,
-    );
-  }
-  const ttlMs = parseDuration(settings.ttl_);
-  const holder: LockHolder = {
-    actor: env.actor,
-    runId: env.runId,
-    since: new Date().toISOString(),
-  };
-  if (env.runUrl !== undefined) holder.runUrl = env.runUrl;
-
-  const result = await store.acquireLock(key, holder, ttlMs);
-  if (!result.ok) {
-    const guidance = settings.onConflict_
-      ? settings.onConflict_(result.holder)
-      : defaultConflictGuidance(key, result.holder);
-    throw new LockConflictError(result.holder, guidance);
-  }
-
-  const token = result.token;
-  // Renew at half the TTL so a long body keeps its short-TTL lock; cleared on
-  // release. The interval is unref'd so it never keeps the process alive.
-  // Renewal is best-effort: `.catch` swallows a rejected renew (store contention
-  // or a transient error) so it can never surface as an unhandled rejection that
-  // crashes the build from a background timer — the lock simply lapses at its
-  // TTL, which is the documented backstop.
-  const heartbeat = setInterval(() => {
-    store.renewLock(key, token, ttlMs).catch(() => {});
-  }, Math.max(1000, Math.floor(ttlMs / 2)));
-  Deno.unrefTimer(heartbeat);
-  return {
-    release: async () => {
-      clearInterval(heartbeat);
-      // Best-effort release for the same reason: a failed release must not turn
-      // an otherwise-succeeded body into a failure. The TTL reclaims the lock.
-      await store.releaseLock(key, token).catch(() => {});
-    },
-  };
-}
-
-/** Resolve a timeout thunk to a JSON-serialisable disposition (default `"fail"`). */
-function resolveDisposition(thunk: OnTimeout | undefined): WaitDisposition {
-  if (thunk === undefined) return "fail";
-  const disposition = thunk();
-  if (disposition === "fail" || disposition === "cancel-run") {
-    return disposition;
-  }
-  return { target: disposition.name_ ?? "?" };
-}
-
-/** The outcome of evaluating a target's `.waitsFor(...)` gate. */
-interface WaitResolution {
-  satisfied: boolean;
-  waitState: WaitState;
-  descriptor: string;
-}
-
-/**
- * Run a target's wait settings lambda, evaluate its trigger against the run's
- * signals, and build the {@link WaitState} to record if it must suspend.
- */
-async function resolveWait(
-  configure: Configure<WaitSettings>,
-  env: RunEnv,
-  name: string,
-): Promise<WaitResolution> {
-  const settings = configure(new WaitSettings());
-  const trigger = settings.trigger_;
-  if (trigger === undefined) {
-    throw new Error(
-      `Target "${name}" .waitsFor(...) set no trigger — call s.on(...).`,
-    );
-  }
-  // The trigger gets the target's durable state handle, so a stateful trigger
-  // (e.g. dispatch-then-poll) can persist correlation state across the
-  // suspend/resume boundary and hand a result to the body.
-  const waitCtx: WaitContext = {
-    state: env.writer ? env.writer.stateHandle(name) : inMemoryStateHandle(),
-    runId: env.runId,
-    target: name,
-  };
-  const satisfied = await trigger.isSatisfied(env.signals, waitCtx);
-  const waitState: WaitState = {
-    trigger: trigger.descriptor,
-    onTimeout: resolveDisposition(settings.onTimeout_),
-  };
-  if (settings.timeout_ !== undefined) {
-    // Preserve the deadline first recorded when this wait suspended: a resume
-    // that re-suspends the same still-unsatisfied gate must not push the
-    // deadline forward, or an hourly `resume --check` cron would reset the
-    // timeout every hour and it would never fire. Only compute a fresh deadline
-    // when there is no prior one (the first suspend).
-    const priorDeadline = env.priorWaits?.get(name)?.deadline;
-    waitState.deadline = priorDeadline ??
-      new Date(Date.now() + parseDuration(settings.timeout_)).toISOString();
-  }
-  return { satisfied, waitState, descriptor: trigger.descriptor };
-}
-
-/**
- * The merged lifecycle: the build's own hooks plus any registered plugins,
- * invoked in order (build first, then each plugin). The run functions call
- * through this so they need not know about plugins.
- */
-interface Lifecycle {
-  start(): Promise<void>;
-  targetStart(name: string): Promise<void>;
-  targetEnd(
-    name: string,
-    status: TargetStatus,
-    durationMs: number,
-  ): Promise<void>;
-  finish(result: BuildResult): Promise<void>;
-  /** Notify plugins of a run-level durable status change (no-op without a store). */
-  runStateChange(record: RunRecord): Promise<void>;
-}
-
-/**
- * Compose a build and its plugins into one {@link Lifecycle}. The run's
- * {@link RunInfo} is bound in, so it enriches every plugin hook without threading
- * it through each call site; the build's own hooks keep their original
- * signatures. Plugin hooks that ignore the extra arguments stay compatible.
- *
- * A plugin is an **observer** — its contract is to report, time, or notify, not
- * to change a target's result — so a throwing plugin hook is caught and reported
- * through `warn`, never allowed to break the run. The build's own hooks are the
- * build's logic and still propagate.
- */
-function makeLifecycle(
-  build: Build,
-  plugins: Plugin[],
-  run: RunInfo,
-  warn: (message: string) => void,
-): Lifecycle {
-  const observe = async (
-    hook: string,
-    call: (p: Plugin) => void | Promise<void>,
-  ): Promise<void> => {
-    for (const p of plugins) {
-      try {
-        await call(p);
-      } catch (error) {
-        warn(
-          `plugin "${p.name ?? "?"}" threw in ${hook}: ${
-            errorMessage(error) ?? "unknown error"
-          } (ignored — plugins observe, they do not change the run)`,
-        );
-      }
-    }
-  };
-  return {
-    async start() {
-      await build.onStart();
-      await observe("onStart", (p) => p.onStart?.(run));
-    },
-    async targetStart(name) {
-      await build.onTargetStart(name);
-      await observe("onTargetStart", (p) => p.onTargetStart?.(name, run));
-    },
-    async targetEnd(name, status, durationMs) {
-      await build.onTargetEnd(name, status);
-      const timing: TargetTiming = { runId: run.runId, durationMs };
-      await observe(
-        "onTargetEnd",
-        (p) => p.onTargetEnd?.(name, status, timing),
-      );
-    },
-    async finish(result) {
-      await build.onFinish(result);
-      await observe("onFinish", (p) => p.onFinish?.(result, run));
-    },
-    async runStateChange(record) {
-      await observe("onRunStateChange", (p) => p.onRunStateChange?.(record));
-    },
-  };
-}
-
-/**
- * Run a target body, applying its {@link TargetBuilder.timeout} and
- * {@link TargetBuilder.retry} settings: each attempt is bounded by the timeout,
- * and a failure is retried (after an optional delay) up to the retry count
- * before the last error propagates.
- */
-async function runBody(t: TargetBuilder, ctx: TargetContext): Promise<void> {
-  const fn = t.fn_;
-  if (fn === undefined) return; // guarded by the caller
-  const attempts = t.retries_ + 1;
-  for (let attempt = 1;; attempt++) {
-    try {
-      await runWithTimeout(() => fn(ctx), t.timeout_);
-      return;
-    } catch (error) {
-      if (attempt >= attempts) throw error;
-      // A cancelled run (Ctrl-C / SIGTERM / an aborted signal) must not burn
-      // further retries: surface the failure now instead of re-running the body.
-      if (ctx.signal.aborted) throw error;
-      if (t.retryDelay_ > 0) await delay(t.retryDelay_);
-    }
-  }
-}
-
-/**
- * Run the target body, and if it fails, hand the failure to each configured
- * {@link TargetBuilder.recoverWith} remediation in turn. When any remediation
- * asks to retry, the body is re-run; this repeats up to
- * {@link TargetBuilder.recoverAttempts} times. The body finally passing resolves
- * normally; otherwise the last failure propagates. A remediation that throws is
- * treated as "could not heal" — it never masks the original build failure.
- */
-async function runBodyWithRecovery(
-  t: TargetBuilder,
-  name: string,
-  globalRecovery: Remediation[],
-  ctx: TargetContext,
-): Promise<void> {
-  try {
-    await runBody(t, ctx);
-    return;
-  } catch (error) {
-    // A target's own remediations run first, then any build-level ones.
-    const remediations = [...t.recoverWith_, ...globalRecovery];
-    if (remediations.length === 0) throw error;
-    let lastError = error;
-    for (let attempt = 1; attempt <= t.recoverAttempts_; attempt++) {
-      // A cancelled run must not trigger (possibly paid) remediations: stop and
-      // let the original failure propagate (falls through to `throw lastError`).
-      if (ctx.signal.aborted) break;
-      let willRetry = false;
-      for (const r of remediations) {
-        try {
-          const result = await r.remediate({
-            target: name,
-            attempt,
-            error: lastError,
-          });
-          if (result.retry) willRetry = true;
-        } catch {
-          // A throwing remediation counts as "could not heal"; keep the build
-          // error intact rather than surfacing the remediation's own failure.
-        }
-      }
-      if (!willRetry) break;
-      try {
-        await runBody(t, ctx);
-        return;
-      } catch (retryError) {
-        lastError = retryError;
-      }
-    }
-    throw lastError;
-  }
-}
-
-/**
- * Run one target: honour its `onlyWhen` conditions and the incremental cache,
- * then (if it must run) open its section, run its body, and report pass/fail.
- * A condition that fails yields `skipped`; an up-to-date target yields `cached`
- * — both unblock dependents without executing the body. With `dryRun`, a target
- * that would run is reported without executing its body or touching the cache.
- */
-async function runTarget(
-  life: Lifecycle,
-  reporter: Reporter,
-  renderer: Renderer,
-  style: Style,
-  t: TargetBuilder,
-  opened: number,
-  cache: BuildCache | undefined,
-  dryRun: boolean,
-  globalRecovery: Remediation[],
-  services: ServiceRegistry,
-  env: RunEnv,
-): Promise<TargetOutcome> {
-  const name = t.name_ ?? "<unnamed>";
-
-  for (const condition of t.onlyWhen_) {
-    if (!(await condition())) return { status: "skipped", ms: 0 };
-  }
-
-  const missing = t.requires_.filter((p) => !p.isSet_());
-  if (missing.length > 0) {
-    const names = missing.map((p) => `"${p.name_ ?? "(unnamed)"}"`).join(", ");
-    const error = new Error(
-      `Target "${name}" requires parameter(s) that are not set: ${names}.`,
-    );
-    openTarget(reporter, renderer, style, name, opened);
-    failTarget(reporter, renderer, style, name, 0, error);
-    return { status: "failed", ms: 0, error };
-  }
-
-  if (dryRun) {
-    openTarget(reporter, renderer, style, name, opened);
-    // A `.dryRunnable()` target with a body runs it with `$` in echo mode
-    // instead of being skipped, to preview the exact commands a real run would
-    // execute. State is in-memory only (a dry run persists nothing), and no lock
-    // or cache is taken.
-    if (t.dryRunnable_ && t.fn_ !== undefined) {
-      const echoState = inMemoryStateHandle();
-      const ctx: TargetContext = {
-        runId: env.runId,
-        target: name,
-        signal: env.signal,
-        state: echoState,
-        stateOf: (t2) => t2 === name ? echoState : inMemoryStateHandle(),
-        signals: env.signals,
-        dryRun: true,
-      };
-      const start = performance.now();
-      try {
-        await withAmbientEcho(
-          (line) => reporter.info(`  $ ${line}`),
-          () => runBody(t, ctx),
-        );
-        const ms = performance.now() - start;
-        passTarget(reporter, renderer, style, name, ms);
-        return { status: "passed", ms };
-      } catch (error) {
-        const ms = performance.now() - start;
-        failTarget(reporter, renderer, style, name, ms, error);
-        return { status: "failed", ms, error };
-      }
-    }
-    for (const line of renderer.targetDryRunFooter(style, name)) {
-      reporter.info(line);
-    }
-    return { status: "passed", ms: 0 };
-  }
-
-  // A service starts a long-lived process and stays up while its dependents
-  // run; the registry stops it during teardown. It has no cacheable body and no
-  // `.executes` — so it is handled before the cache and body paths below.
-  if (t instanceof ServiceBuilder) {
-    openTarget(reporter, renderer, style, name, opened);
-    await life.targetStart(name);
-    void env.writer?.markTargetRunning(name);
-    const start = performance.now();
-    try {
-      services.register(await t.launch_(name));
-      const ms = performance.now() - start;
-      passTarget(reporter, renderer, style, name, ms);
-      return { status: "passed", ms };
-    } catch (error) {
-      const ms = performance.now() - start;
-      failTarget(reporter, renderer, style, name, ms, error);
-      return { status: "failed", ms, error };
-    }
-  }
-
-  // A `.forEach(...)` target fans out into a per-item pipeline of sub-targets,
-  // driven by a nested scheduler. It has no body of its own, so it is handled
-  // before the cache and body paths (like a service).
-  if (t.forEach_ !== undefined) {
-    return await runForEachTarget(
-      t,
-      t.forEach_,
-      life,
-      reporter,
-      renderer,
-      style,
-      opened,
-      cache,
-      dryRun,
-      globalRecovery,
-      services,
-      env,
-    );
-  }
-
-  if (cache !== undefined && await cache.upToDate(t)) {
-    return { status: "cached", ms: 0 };
-  }
-
-  // A `.waitsFor(...)` target is a gate, not a body: if its trigger is already
-  // satisfied it passes (dependents run); otherwise the run suspends here.
-  if (t.waitsFor_ !== undefined) {
-    openTarget(reporter, renderer, style, name, opened);
-    let wait: WaitResolution;
-    try {
-      wait = await resolveWait(t.waitsFor_, env, name);
-    } catch (error) {
-      failTarget(reporter, renderer, style, name, 0, error);
-      return { status: "failed", ms: 0, error };
-    }
-    if (wait.satisfied) {
-      passTarget(reporter, renderer, style, name, 0);
-      return { status: "passed", ms: 0 };
-    }
-    void env.writer?.markTargetWaiting(name, wait.waitState);
-    for (const line of targetWaitFooter(style, name, wait.descriptor)) {
-      reporter.info(line);
-    }
-    return { status: "waiting", ms: 0 };
-  }
-
-  openTarget(reporter, renderer, style, name, opened);
-  await life.targetStart(name);
-  void env.writer?.markTargetRunning(name);
-  const start = performance.now();
-
-  if (!t.fn_) {
-    const error = new Error(
-      `Target "${name}" has no body — call .executes(...) before running.`,
-    );
-    failTarget(reporter, renderer, style, name, 0, error);
-    return { status: "failed", ms: 0, error };
-  }
-
-  // One own-state handle, reused for `stateOf(this target)` so the documented
-  // `stateOf(self) === state` invariant holds even store-less (a fresh
-  // inMemoryStateHandle per call would drop writes).
-  const ownState = env.writer
-    ? env.writer.stateHandle(name)
-    : inMemoryStateHandle();
-  const ctx: TargetContext = {
-    runId: env.runId,
-    target: name,
-    signal: env.signal,
-    state: ownState,
-    stateOf: (t) =>
-      t === name
-        ? ownState
-        : (env.writer ? env.writer.stateHandle(t) : inMemoryStateHandle()),
-    signals: env.signals,
-    dryRun,
-  };
-
-  // Acquire the target's cross-run lock (if any) before the body. A conflict —
-  // or a lock declared with no store — fails the target with the guidance.
-  let lock: HeldLock | null;
-  try {
-    lock = await acquireTargetLock(t, env);
-  } catch (error) {
-    const ms = performance.now() - start;
-    failTarget(reporter, renderer, style, name, ms, error);
-    return { status: "failed", ms, error };
-  }
-
-  try {
-    for (const v of t.validateBefore_) await v.validate({ target: name });
-    await runBodyWithRecovery(t, name, globalRecovery, ctx);
-    for (const v of t.validateAfter_) await v.validate({ target: name });
-    const ms = performance.now() - start;
-    if (cache !== undefined) await cache.record(t);
-    passTarget(reporter, renderer, style, name, ms);
-    return { status: "passed", ms };
-  } catch (error) {
-    const ms = performance.now() - start;
-    failTarget(reporter, renderer, style, name, ms, error);
-    return { status: "failed", ms, error };
-  } finally {
-    // Release on every path — success, failure, cancellation. The TTL is only
-    // the backstop for a killed process.
-    if (lock !== null) await lock.release();
-  }
-}
-
-/**
- * Run a `.forEach(...)` fan-out target: materialise a pipeline of sub-targets
- * per item (named `parent[key].stage`, each stage depending on the previous),
- * then drive them all with the shared {@link runScheduled} machinery — items
- * concurrent up to the configured limit, each item's stages sequential. With
- * `continueOnItemFailure`, sub-targets are lenient so a failed item does not
- * halt its siblings; the fan-out target still fails if any item did. The
- * sub-targets' reports come back as {@link TargetOutcome.children} for the
- * summary, and each is recorded in the run's state under its own name.
- */
-async function runForEachTarget(
-  t: TargetBuilder,
-  spec: ForEachSpec,
-  life: Lifecycle,
-  reporter: Reporter,
-  renderer: Renderer,
-  style: Style,
-  opened: number,
-  cache: BuildCache | undefined,
-  dryRun: boolean,
-  globalRecovery: Remediation[],
-  services: ServiceRegistry,
-  env: RunEnv,
-): Promise<TargetOutcome> {
-  const name = t.name_ ?? "<unnamed>";
-  const settings = spec.configure
-    ? spec.configure(new ForEachSettings())
-    : new ForEachSettings();
-  openTarget(reporter, renderer, style, name, opened);
-  await life.targetStart(name);
-  void env.writer?.markTargetRunning(name);
-  const start = performance.now();
-
-  // Materialise each item's stages into named, chained sub-targets. The items
-  // thunk and factory are user code (they read params and build targets), so a
-  // throw here must fail the fan-out target cleanly — not escape as an uncaught
-  // rejection that crashes the run and leaves the record non-terminal.
-  const order: TargetBuilder[] = [];
-  const predecessors = new Map<TargetBuilder, TargetBuilder[]>();
-  let items: ForEachItem[];
-  try {
-    items = spec.materialize();
-    for (const { key, stages } of items) {
-      let prev: TargetBuilder | undefined;
-      for (const [stage, sub] of Object.entries(stages)) {
-        sub.name_ = `${name}[${key}].${stage}`;
-        // Isolating item failures means a failed stage stays lenient: its
-        // siblings keep running, only this item's later stages are blocked.
-        if (settings.continueOnItemFailure_) sub.proceedAfterFailure_ = true;
-        const deps = prev === undefined ? [] : [prev];
-        if (prev !== undefined) sub.dependsOn_.push(prev);
-        order.push(sub);
-        predecessors.set(sub, deps);
-        prev = sub;
-      }
-    }
-  } catch (error) {
-    const ms = performance.now() - start;
-    failTarget(reporter, renderer, style, name, ms, error);
-    return { status: "failed", ms, error };
-  }
-  reporter.info(
-    items.length === 0
-      ? `${name}: fan-out over 0 items — nothing to run.`
-      : `${name}: fan-out over ${items.length} item(s).`,
-  );
-  const run = await runScheduled(
-    life,
-    order,
-    predecessors,
-    reporter,
-    renderer,
-    style,
-    new Set(),
-    settings.concurrency_ ?? cpuCount(),
-    () => true, // items and stages overlap; predecessors enforce per-item order
-    cache,
-    dryRun,
-    globalRecovery,
-    services,
-    env,
-  );
-  const ms = performance.now() - start;
-  if (run.aborted) {
-    const failed = run.reports.filter((r) => r.status === "failed").length;
-    const error = run.failure ??
-      new Error(`${name}: ${failed} sub-target(s) failed.`);
-    failTarget(reporter, renderer, style, name, ms, error);
-    return { status: "failed", ms, error, children: run.reports };
-  }
-  passTarget(reporter, renderer, style, name, ms);
-  return { status: "passed", ms, children: run.reports };
-}
-
-/** Resolve the concurrency limit; 1 means sequential. */
-function resolveConcurrency(option: boolean | number | undefined): number {
-  if (option === undefined || option === false) return 1;
-  if (option === true) return cpuCount();
-  return option > 1 ? Math.floor(option) : 1;
-}
-
-/** The host's CPU count, used as the default parallel/batch concurrency. */
-function cpuCount(): number {
-  const cpus = navigator.hardwareConcurrency;
-  return cpus > 0 ? cpus : 4;
-}
-
-/** A reporter that buffers lines so a target's block can flush atomically. */
-function bufferReporter(): {
-  reporter: Reporter;
-  flush: (to: Reporter) => void;
-} {
-  const lines: Array<{ error: boolean; text: string }> = [];
-  return {
-    reporter: {
-      info: (text) => void lines.push({ error: false, text }),
-      error: (text) => void lines.push({ error: true, text }),
-    },
-    flush: (to) => {
-      for (const line of lines) {
-        if (line.error) to.error(line.text);
-        else to.info(line.text);
-      }
-    },
-  };
-}
-
-/** Sequentially run the plan, aborting (and skipping the rest) on first failure. */
-async function runSequential(
-  life: Lifecycle,
-  order: TargetBuilder[],
-  reporter: Reporter,
-  renderer: Renderer,
-  style: Style,
-  skip: Set<string>,
-  cache: BuildCache | undefined,
-  dryRun: boolean,
-  globalRecovery: Remediation[],
-  services: ServiceRegistry,
-  env: RunEnv,
-): Promise<RunOutcome> {
-  const reports: TargetReport[] = [];
-  const executed: string[] = [];
-  let failure: unknown;
-  let aborted = false;
-  let opened = 0;
-
-  for (const t of order) {
-    const name = t.name_ ?? "<unnamed>";
-    if (skip.has(name) || aborted) {
-      reports.push({ name, status: "skipped", ms: 0 });
-      void env.writer?.markTargetSettled(name, "skipped");
-      continue;
-    }
-    let outcome: TargetOutcome;
-    try {
-      outcome = await runTarget(
-        life,
-        reporter,
-        renderer,
-        style,
-        t,
-        opened,
-        cache,
-        dryRun,
-        globalRecovery,
-        services,
-        env,
-      );
-      await life.targetEnd(name, outcome.status, outcome.ms);
-    } catch (error) {
-      // A reject from outside runTarget's own try/catch (an `onlyWhen`/`cacheKey`
-      // thunk, a lifecycle hook, or `life.targetEnd`) becomes a failed target so
-      // the run finalizes here instead of rejecting out of `execute()` and
-      // stranding the record `running`.
-      failTarget(reporter, renderer, style, name, 0, error);
-      outcome = { status: "failed", ms: 0, error };
-    }
-    void env.writer?.markTargetSettled(
-      name,
-      outcome.status,
-      errorMessage(outcome.error),
-    );
-    if (outcome.status === "passed" || outcome.status === "failed") opened++;
-    reports.push({ name, status: outcome.status, ms: outcome.ms });
-    // A fan-out target's sub-targets appear as their own rows beneath it.
-    if (outcome.children !== undefined) reports.push(...outcome.children);
-    if (outcome.status === "passed") executed.push(name);
-    else if (outcome.status === "failed") {
-      failure = outcome.error;
-      aborted = true;
-    }
-  }
-  return { reports, executed, failure, aborted, suspended: false };
-}
-
-/**
- * Run the plan with up to `limit` targets in flight, respecting dependencies.
- * `canOverlap` decides which ready targets may run at the same time: with
- * global parallelism it is always true; otherwise only members of the same
- * {@link group} overlap, keeping ungrouped targets serialized.
- *
- * Each target's framework output is buffered and flushed as a contiguous block
- * on completion, so concurrent runs don't interleave their banners. A failure
- * stops new launches; in-flight targets settle and the rest are skipped.
- */
-async function runScheduled(
-  life: Lifecycle,
-  order: TargetBuilder[],
-  predecessors: Map<TargetBuilder, TargetBuilder[]>,
-  reporter: Reporter,
-  renderer: Renderer,
-  style: Style,
-  skip: Set<string>,
-  limit: number,
-  canOverlap: (a: TargetBuilder, b: TargetBuilder) => boolean,
-  cache: BuildCache | undefined,
-  dryRun: boolean,
-  globalRecovery: Remediation[],
-  services: ServiceRegistry,
-  env: RunEnv,
-): Promise<RunOutcome> {
-  const outcomes = new Map<TargetBuilder, TargetOutcome>();
-  const done = new Set<TargetBuilder>(); // passed/cached/skipped → unblocks dependents
-  const started = new Set<TargetBuilder>();
-  const runningSet = new Set<TargetBuilder>();
-  let failure: unknown;
-  let anyFailed = false; // a failure occurred → the build fails
-  let anyWaiting = false; // a `.waitsFor(...)` gate parked → the run suspends
-  let halted = false; // a non-lenient failure → stop launching new targets
-  let flushed = 0;
-
-  // `--skip` targets, and targets already succeeded on a resumed run, count as
-  // completed so their dependents can still run.
-  for (const t of order) {
-    const name = t.name_ ?? "<unnamed>";
-    if (skip.has(name)) {
-      outcomes.set(t, { status: "skipped", ms: 0 });
-      done.add(t);
-      started.add(t);
-      void env.writer?.markTargetSettled(name, "skipped");
-    } else if (env.done?.has(name)) {
-      // Succeeded in the prior (suspended) run: unblock dependents, don't re-run,
-      // and leave its recorded `succeeded` untouched.
-      outcomes.set(t, { status: "cached", ms: 0 });
-      done.add(t);
-      started.add(t);
-    }
-  }
-
-  const ready = (t: TargetBuilder): boolean =>
-    (predecessors.get(t) ?? []).every((p) => done.has(p));
-  const overlaps = (t: TargetBuilder): boolean =>
-    [...runningSet].every((r) => canOverlap(t, r));
-
-  await new Promise<void>((resolve) => {
-    const pump = () => {
-      for (const t of order) {
-        if (runningSet.size >= limit) break;
-        if (started.has(t) || !ready(t) || !overlaps(t)) continue;
-        // After a fatal failure, stop launching — except `always` targets,
-        // which run for cleanup even when the build is failing.
-        if (halted && !t.always_) continue;
-        started.add(t);
-        runningSet.add(t);
-        const buffer = bufferReporter();
-        runTarget(
-          life,
-          buffer.reporter,
-          renderer,
-          style,
-          t,
-          flushed,
-          cache,
-          dryRun,
-          globalRecovery,
-          services,
-          env,
-        )
-          .then(
-            async (outcome) => {
-              await life.targetEnd(
-                t.name_ ?? "<unnamed>",
-                outcome.status,
-                outcome.ms,
-              );
-              // A waiting gate already recorded its `waitingFor` via
-              // markTargetWaiting; settling it here would clobber that.
-              if (outcome.status !== "waiting") {
-                void env.writer?.markTargetSettled(
-                  t.name_ ?? "<unnamed>",
-                  outcome.status,
-                  errorMessage(outcome.error),
-                );
-              }
-              // An executed target (or a parked wait) prints a block worth
-              // separating.
-              const printed = outcome.status === "passed" ||
-                outcome.status === "failed" || outcome.status === "waiting";
-              if (printed) {
-                if (!style.github && flushed > 0) reporter.info("");
-                buffer.flush(reporter);
-                flushed++;
-              }
-              outcomes.set(t, outcome);
-              runningSet.delete(t);
-              if (outcome.status === "failed") {
-                anyFailed = true;
-                failure ??= outcome.error;
-                // A lenient failure lets independent targets keep going; its
-                // own dependents stay blocked (never added to `done`).
-                if (!t.proceedAfterFailure_) halted = true;
-              } else if (outcome.status === "waiting") {
-                // The run suspends here: this target's dependents stay blocked
-                // (never `done`), but independent branches run to completion.
-                anyWaiting = true;
-              } else {
-                done.add(t); // passed, cached, or condition-skipped
-              }
-              pump();
-            },
-          )
-          .catch((error: unknown) => {
-            // `runTarget` handles a thrown body itself, but paths outside its
-            // try/catch can still reject — an `onlyWhen`/`cacheKey` thunk, a
-            // lifecycle hook, or `life.targetEnd` in the fulfilment handler
-            // above. Route any such rejection into the normal failure path so
-            // the scheduler settles (and finalizes the record) instead of
-            // hanging with the target stuck in `runningSet`, or dying on an
-            // unhandled rejection.
-            //
-            // Free the slot and record the failure BEFORE emitting output, so
-            // even a throwing reporter (already made best-effort by
-            // safeReporter) can never leave `t` in `runningSet` — which would
-            // wedge the scheduler's completion `Promise` forever.
-            const targetName = t.name_ ?? "<unnamed>";
-            outcomes.set(t, { status: "failed", ms: 0, error });
-            runningSet.delete(t);
-            anyFailed = true;
-            failure ??= error;
-            if (!t.proceedAfterFailure_) halted = true;
-            failTarget(reporter, renderer, style, targetName, 0, error);
-            void env.writer?.markTargetSettled(
-              targetName,
-              "failed",
-              errorMessage(error),
-            );
-            pump();
-          });
-      }
-      if (runningSet.size === 0) {
-        for (const t of order) {
-          if (!started.has(t)) {
-            outcomes.set(t, { status: "skipped", ms: 0 });
-            started.add(t);
-            // When the run *suspends* (a wait parked and nothing failed),
-            // targets blocked behind the wait are left `pending` so a resume
-            // runs them. A run that also failed will never resume, so settle
-            // them `skipped` — otherwise they linger `pending` inside a terminal
-            // `failed` record that no resume sweep reaches (F8).
-            if (!anyWaiting || anyFailed) {
-              void env.writer?.markTargetSettled(
-                t.name_ ?? "<unnamed>",
-                "skipped",
-              );
-            }
-          }
-        }
-        resolve();
-      }
-    };
-    pump();
-  });
-
-  // A failed run does not suspend, so any target parked at a `.waitsFor(...)`
-  // gate — recorded `waiting` and normally left for a resume — will never be
-  // resumed. Settle those rows to a terminal `skipped` so the `failed` record
-  // has no permanently-`waiting` target that `runs show` and the resume sweep
-  // would otherwise see forever (F8).
-  if (anyFailed && anyWaiting) {
-    for (const [t, outcome] of outcomes) {
-      if (outcome.status === "waiting") {
-        outcomes.set(t, { status: "skipped", ms: outcome.ms });
-        void env.writer?.markTargetSettled(t.name_ ?? "<unnamed>", "skipped");
-      }
-    }
-  }
-
-  const reports: TargetReport[] = [];
-  const executed: string[] = [];
-  for (const t of order) {
-    const name = t.name_ ?? "<unnamed>";
-    const outcome = outcomes.get(t) ?? { status: "skipped", ms: 0 };
-    reports.push({ name, status: outcome.status, ms: outcome.ms });
-    // A fan-out target's sub-targets appear as their own rows beneath it.
-    if (outcome.children !== undefined) reports.push(...outcome.children);
-    if (outcome.status === "passed") executed.push(name);
-  }
-  return {
-    reports,
-    executed,
-    failure,
-    aborted: anyFailed,
-    suspended: anyWaiting && !anyFailed,
-  };
-}
-
 /**
  * Execute the requested target and its transitive dependencies.
  *
@@ -1677,21 +607,20 @@ export async function execute(
   // Services started during the run are held here and torn down in reverse
   // order once it finishes — in a `finally`, so a failure never leaks a process.
   const services = new ServiceRegistry();
+  const ctx: RunContext = {
+    life,
+    reporter,
+    renderer,
+    style,
+    cache,
+    dryRun,
+    globalRecovery,
+    services,
+    env,
+  };
   const runPlan = (): Promise<RunOutcome> => {
     if (!globalParallel && !grouped && !scheduled) {
-      return runSequential(
-        life,
-        order,
-        reporter,
-        renderer,
-        style,
-        skip,
-        cache,
-        dryRun,
-        globalRecovery,
-        services,
-        env,
-      );
+      return runSequential(ctx, order, skip);
     }
     // With `--parallel`, anything independent may overlap up to `limit`.
     // Otherwise only same-group members overlap (the rest stay serialized),
@@ -1702,20 +631,12 @@ export async function execute(
       : (a: TargetBuilder, b: TargetBuilder) =>
         a.group_ !== undefined && a.group_ === b.group_;
     return runScheduled(
-      life,
+      ctx,
       order,
       predecessors,
-      reporter,
-      renderer,
-      style,
       skip,
       effectiveLimit,
       canOverlap,
-      cache,
-      dryRun,
-      globalRecovery,
-      services,
-      env,
     );
   };
